Admin Bar, Explore search, Accout, Workout modeling

+++ server/controllers/workoutsController.js
@@ -2,25 +2,33 @@ const Workout = require('../models/Workout');
 const User = require('../models/User');
 
 const { format } = require('date-fns');
+const { ObjectId } = require('mongodb');
 
 const fetchAll = async (req, res) => {
   console.log('workouts fetchAll');
+  const userId = req?.cookies.userId;
+  if (!userId)
+    return res.status(500).json({ message: 'no userId found in request' });
 
   const Workouts = await Workout.find().exec();
   if (!Workouts) {
-    return res.status(500).json({ message: 'err while workouts fetchAll' });
+    return res.status(204).send({ message: 'No workouts found' });
   } else {
     return res.status(200).send({ Workouts }).end();
   }
 };
 
 const fetchById = async (req, res) => {
+  console.log('workout fetchById');
+  const userId = req?.cookies.userId;
+  if (!userId)
+    return res.status(500).json({ message: 'no userId found in request' });
+
   const workoutId = req.params.workoutId;
   if (!workoutId)
     return res.status(400).json({ message: 'workout id is required' });
 
   const foundWorkout = await Workout.findById(workoutId).exec();
-  console.log(foundWorkout);
   if (!foundWorkout) {
     return res
       .status(400)
@@ -41,9 +49,10 @@ const fetchByUser = async (req, res) => {
   if (!foundUser)
     return res.status(500).json({ message: 'request userId dont exist id DB' });
 
-  const username = foundUser.username;
+  // request is OK.
 
-  const Workouts = await Workout.find({ user: username }).exec();
+  const id = foundUser._id;
+  const Workouts = await Workout.find({ user_id: id }).exec();
   if (!Workouts) {
     return res.status(500).json({ message: 'err while workouts fetchAll' });
   } else {
@@ -60,6 +69,12 @@ const deleteById = async (req, res) => {
   if (!workoutId)
     return res.status(400).json({ message: 'workout id is required' });
 
+  const loggedInUser = await User.findById(userId);
+  if (!loggedInUser)
+    return res.status(500).json({ message: 'request userId dont exist id DB' });
+
+  // request is OK.
+
   const foundWorkout = await Workout.findOneAndDelete({
     _id: workoutId,
   }).exec();
@@ -69,14 +84,9 @@ const deleteById = async (req, res) => {
       .json({ message: 'Workout was not found - Bad ID' })
       .send();
 
-  // return Workouts for the user for re-rendering...
-  const foundUser = await User.findById(userId);
-  if (!foundUser)
-    return res.status(500).json({ message: 'request userId dont exist id DB' });
-
-  const name = foundUser.username;
-
-  const Workouts = await Workout.find({ user: name }).exec();
+  // for re-rendering, return updated Workouts for the user.
+  const id = foundUser._id;
+  const Workouts = await Workout.find({ user_id: id }).exec();
   if (!Workouts) {
     return res.status(500).json({ message: 'err while workouts fetchAll' });
   } else {
@@ -93,24 +103,30 @@ const addWorkout = async (req, res) => {
   if (!workout) {
     return res.status(400).json({ message: 'bad add Workout request' }).end();
   }
+
+  //request is OK.
+
+  //find user in DB.
   const foundUser = await User.findById(userId);
   if (!foundUser)
     return res.status(500).json({ message: 'request userId dont exist id DB' });
 
-  const username = foundUser.username;
+  const { firstname, lastname } = foundUser.name;
+
   console.log(workout);
   const { title, exercises } = workout;
 
   await Workout.create({
-    user: username,
+    user_id: new ObjectId(userId),
+    author: { firstname, lastname },
     title: title,
     lastUpdated: format(new Date(), 'dd/MM/YYY pp'),
     exercises: exercises,
   });
 
-  // return Workouts for the user for re-rendering...
-
-  const Workouts = await Workout.find({ user: username }).exec();
+  // for re-rendering, return updated Workouts for the user.
+  const id = foundUser._id;
+  const Workouts = await Workout.find({ user_id: id }).exec();
   if (!Workouts) {
     return res.status(500).json({ message: 'err while workouts fetchAll' });
   } else {
@@ -119,34 +135,48 @@ const addWorkout = async (req, res) => {
 };
 
 const updateById = async (req, res) => {
+  console.log('workout updateById');
   const userId = req?.cookies.userId;
+  const { workout } = req.body;
+  const workoutId = req.params.workoutId;
+
   if (!userId)
     return res.status(500).json({ message: 'no userId found in request' });
 
-  const { workout } = req.body;
   if (!workout) {
-    return res.status(400).json({ message: 'bad add Workout request' }).end();
+    return res.status(400).json({ message: 'bad workout body request' }).end();
   }
-  const workoutId = req.params.workoutId;
   if (!workoutId)
     return res.status(400).json({ message: 'bad workout id request' });
+
+  // request is ok.
+
+  if (workout.user_id !== userId) {
+    console.log('user trying to update others workout');
+    return res.status(400).json({ message: 'workout dont belong to user' });
+  }
+
+  // find user in DB.
   const foundUser = await User.findById(userId).exec();
   if (!foundUser)
     return res
       .status(500)
       .json({ message: `request userId don't exist in DB` });
 
+  // update workout in DB.
   const result = await Workout.updateOne(
     { _id: workoutId },
     {
       title: workout.title,
       exercises: workout.exercises,
+      lastUpdated: format(new Date(), 'dd/MM/YYY pp'),
     }
   ).exec();
   console.log(result);
 
-  // for re-rendering.
-  const Workouts = await Workout.find({ user: foundUser.username }).exec();
+  // for re-rendering, return updated Workouts for the user.
+  const id = foundUser._id;
+  const Workouts = await Workout.find({ user_id: id }).exec();
   if (!Workouts) {
     return res.status(500).json({ message: 'err while workouts fetchAll' });
   } else {
@@ -154,4 +184,11 @@ const updateById = async (req, res) => {
   }
 };
 
-module.exports = { fetchAll, fetchById, addWorkout, deleteById, fetchByUser };
+module.exports = {
+  fetchAll,
+  fetchById,
+  addWorkout,
+  deleteById,
+  fetchByUser,
+  updateById,
+};
